// src/proto/jon/jon_shared_data_lrf.ts
export interface JonGuiDataLrf {
  isScanning: boolean;
  isMeasuring: boolean;
  measureId: number;
  target: JonGuiDataTarget | undefined;
  meteo: JonGuiDataMeteo | undefined;
}

export interface JonGuiDataTarget {
  timestamp: number;
  targetLongitude: number;
  targetLatitude: number;
  targetAltitude: number;
  observerLongitude: number;
  observerLatitude: number;
  observerAltitude: number;
  observerAzimuth: number;
  observerElevation: number;
  observerBank: number;
  /** Decemetres. In 2D, the distance between the observer and the target. */
  distance2d: number;
  /** Decemetres (3D). Accounting for the altitude difference. */
  distance3b: number;
  observerFixType: JonGuiDataGpsFixType;
  rangeA: number;
  rangeB: number;
  rangeC: number;
  /** Radius of the displayed area based on precision (measurement conditions). */
  radius: number;
  sessionId: number;
  targetId: number;
  /** RGB color representation */
  targetColor:
    | RgbColor
    | undefined;
  /** Tank, soldier... */
  type: number;
}

export interface RgbColor {
  /** 0-255 */
  red: number;
  /** 0-255 */
  green: number;
  /** 0-255 */
  blue: number;
}

function createBaseJonGuiDataLrf(): JonGuiDataLrf {
  return { isScanning: false, isMeasuring: false, measureId: 0, target: undefined, meteo: undefined };
}

export const JonGuiDataLrf = {
  encode(message: JonGuiDataLrf, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.isScanning !== false) {
      writer.uint32(8).bool(message.isScanning);
    }
    if (message.isMeasuring !== false) {
      writer.uint32(16).bool(message.isMeasuring);
    }
    if (message.measureId !== 0) {
      writer.uint32(24).int32(message.measureId);
    }
    if (message.target !== undefined) {
      JonGuiDataTarget.encode(message.target, writer.uint32(34).fork()).ldelim();
    }
    if (message.meteo !== undefined) {
      JonGuiDataMeteo.encode(message.meteo, writer.uint32(42).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): JonGuiDataLrf {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseJonGuiDataLrf();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.isScanning = reader.bool();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.isMeasuring = reader.bool();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.measureId = reader.int32();
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          message.target = JonGuiDataTarget.decode(reader, reader.uint32());
          continue;
        case 5:
          if (tag !== 42) {
            break;
          }

          message.meteo = JonGuiDataMeteo.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): JonGuiDataLrf {
    return {
      isScanning: isSet(object.isScanning) ? globalThis.Boolean(object.isScanning) : false,
      isMeasuring: isSet(object.isMeasuring) ? globalThis.Boolean(object.isMeasuring) : false,
      measureId: isSet(object.measureId) ? globalThis.Number(object.measureId) : 0,
      target: isSet(object.target) ? JonGuiDataTarget.fromJSON(object.target) : undefined,
      meteo: isSet(object.meteo) ? JonGuiDataMeteo.fromJSON(object.meteo) : undefined,
    };
  },

  toJSON(message: JonGuiDataLrf): unknown {
    const obj: any = {};
    if (message.isScanning !== false) {
      obj.isScanning = message.isScanning;
    }
    if (message.isMeasuring !== false) {
      obj.isMeasuring = message.isMeasuring;
    }
    if (message.measureId !== 0) {
      obj.measureId = Math.round(message.measureId);
    }
    if (message.target !== undefined) {
      obj.target = JonGuiDataTarget.toJSON(message.target);
    }
    if (message.meteo !== undefined) {
      obj.meteo = JonGuiDataMeteo.toJSON(message.meteo);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<JonGuiDataLrf>, I>>(base?: I): JonGuiDataLrf {
    return JonGuiDataLrf.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<JonGuiDataLrf>, I>>(object: I): JonGuiDataLrf {
    const message = createBaseJonGuiDataLrf();
    message.isScanning = object.isScanning ?? false;
    message.isMeasuring = object.isMeasuring ?? false;
    message.measureId = object.measureId ?? 0;
    message.target = (object.target !== undefined && object.target !== null)
      ? JonGuiDataTarget.fromPartial(object.target)
      : undefined;
    message.meteo = (object.meteo !== undefined && object.meteo !== null)
      ? JonGuiDataMeteo.fromPartial(object.meteo)
      : undefined;
    return message;
  },
};
// ...
export const RgbColor = {
  encode(message: RgbColor, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.red !== 0) {
      writer.uint32(8).uint32(message.red);
    }
    if (message.green !== 0) {
      writer.uint32(16).uint32(message.green);
    }
    if (message.blue !== 0) {
      writer.uint32(24).uint32(message.blue);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): RgbColor {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRgbColor();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.red = reader.uint32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.green = reader.uint32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.blue = reader.uint32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RgbColor {
    return {
      red: isSet(object.red) ? globalThis.Number(object.red) : 0,
      green: isSet(object.green) ? globalThis.Number(object.green) : 0,
      blue: isSet(object.blue) ? globalThis.Number(object.blue) : 0,
    };
  },

  toJSON(message: RgbColor): unknown {
    const obj: any = {};
    if (message.red !== 0) {
      obj.red = Math.round(message.red);
    }
    if (message.green !== 0) {
      obj.green = Math.round(message.green);
    }
    if (message.blue !== 0) {
      obj.blue = Math.round(message.blue);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RgbColor>, I>>(base?: I): RgbColor {
    return RgbColor.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RgbColor>, I>>(object: I): RgbColor {
    const message = createBaseRgbColor();
    message.red = object.red ?? 0;
    message.green = object.green ?? 0;
    message.blue = object.blue ?? 0;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };